requirements.txt
numpy>=1.21.0
python-Levenshtein>=0.12.2
matplotlib>=3.4.0
pandas>=1.3.0
tqdm>=4.62.0